+++ packages/headless/src/features/facets/range-facets/numeric-facet-set/numeric-facet-set-slice.test.ts
@@ -17,8 +17,9 @@ import * as RangeFacetReducers from '../generic/range-facet-reducers';
 import * as FacetReducers from '../../generic/facet-reducer-helpers';
 import {executeSearch} from '../../../search/search-actions';
 import {buildMockSearch} from '../../../../test/mock-search';
-import {logGenericSearchEvent} from '../../../analytics/analytics-actions';
+import {logSearchEvent} from '../../../analytics/analytics-actions';
 import {numericFacetSetReducer} from './numeric-facet-set-slice';
+import {deselectAllFacets} from '../../generic/facet-actions';
 
 describe('numeric-facet-set slice', () => {
   let state: NumericFacetSetState;
@@ -96,6 +97,19 @@ describe('numeric-facet-set slice', () => {
     ).toHaveBeenCalledTimes(1);
   });
 
+  it('dispatching #deselectAllFacets calls #handleFacetDeselectAll for every numeric facet', () => {
+    jest.spyOn(RangeFacetReducers, 'handleRangeFacetDeselectAll');
+
+    state['1'] = buildMockNumericFacetRequest();
+    state['2'] = buildMockNumericFacetRequest();
+    state['3'] = buildMockNumericFacetRequest();
+    numericFacetSetReducer(state, deselectAllFacets);
+
+    expect(
+      RangeFacetReducers.handleRangeFacetDeselectAll
+    ).toHaveBeenCalledTimes(4);
+  });
+
   it('#updateNumericFacetSortCriterion calls #handleFacetSortCriterionUpdate', () => {
     jest.spyOn(FacetReducers, 'handleFacetSortCriterionUpdate');
 
@@ -116,7 +130,7 @@ describe('numeric-facet-set slice', () => {
     const search = buildMockSearch();
     numericFacetSetReducer(
       state,
-      executeSearch.fulfilled(search, '', logGenericSearchEvent({evt: 'foo'}))
+      executeSearch.fulfilled(search, '', logSearchEvent({evt: 'foo'}))
     );
 
     expect(